# Rewrites: traverse tree once, applying all rules