… scalafmt-core/shared/src/main/scala/org/scalafmt/rewrite/RewriteTrailingCommas.scala
… scalafmt-core/shared/src/main/scala/org/scalafmt/rewrite/SortImports.scala
… scalafmt-core/shared/src/main/scala/org/scalafmt/rewrite/SortModifiers.scala