Also, turn rewrites into sessions with the context instance.