@olafurpg 2020-02-06
This is an interesting change! Do you have any numbers to motivate this optimization?