@kitbellew 2020-02-06
> This is an interesting change! Do you have any numbers to motivate this optimization?

@olafurpg i usually compare run times on `scala-repos` when the run results in no changes (so no time is spent on showing diffs). this change seemed to produce a minor, barely perceptible speedup (i used all rewrites except AvoidInfix) but this approach is quite random and not very convincing.

can you help me with setting up of your benchmarking tools? i tried `./bin/run-benchmarks.sh` but it asks to run `sbt test` which fails with `(dynamic / Test / executeTests) java.lang.NoClassDefFoundError: org/scalatest/FunSuiteLike`. i am not very familiar with `sbt` (we use maven) so i don't know how to fix it.