@olafurpg 2020-02-06
The following command will run a benchmark on a single file with ~4k lines of code with some rewrites enabled, 10 warmup iterations and 10 benchmark iterations.

```
benchmarks/jmh:run -i 10 -wi 10 -f1 -t1 org.scalafmt.benchmarks.Micro.ExtraLarge.scalafmt_rewrite
```

You can tweak the `scalafmt_rewrite` method to enable more rules.