@olafurpg 2020-02-06
It prints something like this for me
```
[info] # Run progress: 0.00% complete, ETA 00:00:20
[info] # Fork: 1 of 1
[info] # Warmup Iteration   1: 3349.428 ms/op
[info] # Warmup Iteration   2: 1385.353 ms/op
[info] # Warmup Iteration   3: 1323.863 ms/op
[info] # Warmup Iteration   4: 1314.509 ms/op
[info] # Warmup Iteration   5: 1180.080 ms/op
[info] # Warmup Iteration   6: 1119.079 ms/op
[info] # Warmup Iteration   7: 1079.098 ms/op
[info] # Warmup Iteration   8: 1043.373 ms/op
[info] # Warmup Iteration   9: 1031.648 ms/op
[info] # Warmup Iteration  10: 1026.872 ms/op
[info] Iteration   1: 958.440 ms/op
[info] Iteration   2: 955.910 ms/op
[info] Iteration   3: 924.383 ms/op
[info] Iteration   4: 984.063 ms/op
[info] Iteration   5: 1022.945 ms/op
[info] Iteration   6: 989.306 ms/op
[info] Iteration   7: 983.143 ms/op
[info] Iteration   8: 983.382 ms/op
[info] Iteration   9: 965.896 ms/op
[info] Iteration  10: 1016.920 ms/op
[info] Result "org.scalafmt.benchmarks.Micro.ExtraLarge.scalafmt_rewrite":
[info]   978.439 ±(99.9%) 44.063 ms/op [Average]
[info]   (min, avg, max) = (924.383, 978.439, 1022.945), stdev = 29.145
[info]   CI (99.9%): [934.376, 1022.502] (assumes normal distribution)
[info] # Run complete. Total time: 00:00:32
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                          Mode  Cnt    Score    Error  Units
[info] Micro.ExtraLarge.scalafmt_rewrite  avgt   10  978.439 ± 44.063  ms/op
```