@kitbellew 2020-02-07
@olafurpg ok, so the results of the benchmark. i set in `FormatBenchmark.scala` the following:
```
  rules = Seq(SortImports, RedundantBraces, PreferCurlyFors, SortModifiers)
```

then ran your command through `sbt` before and after this commit.
* before:
```
[info]   880.639 ±(99.9%) 49.038 ms/op [Average]
[info]   (min, avg, max) = (824.018, 880.639, 941.453), stdev = 32.436
[info]   CI (99.9%): [831.601, 929.677] (assumes normal distribution)
```
* after:
```
[info]   844.943 ±(99.9%) 27.216 ms/op [Average]
[info]   (min, avg, max) = (818.603, 844.943, 869.071), stdev = 18.002
[info]   CI (99.9%): [817.726, 872.159] (assumes normal distribution)
```

then I ran again, in the opposite order:
* after:
```
[info]   741.733 ±(99.9%) 29.887 ms/op [Average]
[info]   (min, avg, max) = (716.586, 741.733, 775.446), stdev = 19.768
[info]   CI (99.9%): [711.846, 771.620] (assumes normal distribution)
```
* before:
```
[info]   761.939 ±(99.9%) 34.677 ms/op [Average]
[info]   (min, avg, max) = (737.170, 761.939, 816.624), stdev = 22.937
[info]   CI (99.9%): [727.262, 796.616] (assumes normal distribution)
```

_some_ improvement (3-5%), although nothing to write home about.