@kitbellew 2020-02-08
@olafurpg @poslegm what do you think, should we proceed or abandon? the performance improvement is slight but positive in both runs.