@olafurpg 2020-02-08
I’m torn. There is a tradeoff since the new design is more complicated IMO. It’s easy to explain “every rule traverses the entire tree”, and I’m not sure if the speed ups are significant enough to justify the introduced complexity. What do you think?

I haven’t worked on these rules much for a while so I’m ok going ahead with this change if you want to.