@kitbellew 2020-02-08
@olafurpg i don't feel too strongly but surprised about the complexity. traverse tree, rewrite each node using a few rules... doesn't sound much different :-)